requirements.txt
fastapi==0.121.2
pydantic==2.12.4
python-dotenv==1.2.1
python_bcrypt==0.3.2
Requests==2.32.5
